Enforce prettier style guide
VSCode formatters (and other Prettier-supported IDEs, e.g JetBrains) will now enforce coding guidelines provided by prettier to keep future development consistent. A few overrides were included:

- Tab width set to 4, as this is what the project was using already
- Trailing commas for brackets

This commit would conflict with people currently working on the code and can pollute the commit history--but as it's the start of the semester this is likely the best time to do it (the earlier the better).